requirements.txt

typing==3.10.0.2
python-dotenv==0.19.2
fastapi==0.70.0
pydantic==1.8.2
